+++ test/test_oneview_enclosure.py
@@ -1,7 +1,7 @@
 #!/usr/bin/python
 # -*- coding: utf-8 -*-
 ###
-# Copyright (2016-2017) Hewlett Packard Enterprise Development LP
+# Copyright (2016-2019) Hewlett Packard Enterprise Development LP
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # You may not use this file except in compliance with the License.
@@ -256,10 +256,12 @@ PARAMS_FOR_INTERCONNECT_BAY_POWER_STATE_E_FUSE = """
 @pytest.mark.resource(TestEnclosureModule='enclosures')
 class TestEnclosureModule(OneViewBaseTest):
     def test_should_create_new_enclosure(self):
-        self.resource.get_by.return_value = []
-        self.resource.get_all.return_value = []
-        self.resource.add.return_value = ENCLOSURE_FROM_ONEVIEW
-        self.resource.patch.return_value = ENCLOSURE_FROM_ONEVIEW
+        self.resource.get_by_name.return_value = []
+        self.resource.get_by_hostname.return_value = []
+
+        self.resource.data = ENCLOSURE_FROM_ONEVIEW
+        self.resource.add.return_value = self.resource
+        self.resource.patch.return_value = self.resource
 
         self.mock_ansible_module.params = PARAMS_FOR_PRESENT
 
@@ -271,9 +273,20 @@ class TestEnclosureModule(OneViewBaseTest):
             ansible_facts=dict(enclosure=ENCLOSURE_FROM_ONEVIEW)
         )
 
+    def test_should_fail_create_new_enclosure(self):
+        self.resource.get_by_name.return_value = []
+        self.resource.get_by_hostname.return_value = []
+        self.resource.add.return_value = []
+        self.mock_ansible_module.params = PARAMS_FOR_PRESENT
+
+        EnclosureModule().run()
+
+        self.mock_ansible_module.fail_json.assert_called_once_with(exception=mock.ANY,
+                                                                   msg=EnclosureModule.MSG_ENCLOSURE_REQUIRED_FIELDS)
+
     def test_should_not_update_when_no_changes_by_primary_ip_key(self):
-        self.resource.get_by.return_value = [ENCLOSURE_FROM_ONEVIEW]
-        self.resource.get_all.return_value = ALL_ENCLOSURES
+        self.resource.data = ENCLOSURE_FROM_ONEVIEW
+        self.resource.get_by_hostname.return_value = self.resource
 
         self.mock_ansible_module.params = PARAMS_FOR_PRESENT
 
@@ -286,8 +299,8 @@ class TestEnclosureModule(OneViewBaseTest):
         )
 
     def test_should_not_update_when_no_changes_by_standby_ip_key(self):
-        self.resource.get_by.return_value = [ENCLOSURE_FROM_ONEVIEW]
-        self.resource.get_all.return_value = ALL_ENCLOSURES
+        self.resource.data = ENCLOSURE_FROM_ONEVIEW
+        self.resource.get_by_hostname.return_value = self.resource
 
         params = deepcopy(PARAMS_FOR_PRESENT)
         params['data']['hostname'] = STANDBY_IP_ADDRESS
@@ -302,8 +315,8 @@ class TestEnclosureModule(OneViewBaseTest):
         )
 
     def test_should_not_update_when_no_changes_by_name_key(self):
-        self.resource.get_by.return_value = [ENCLOSURE_FROM_ONEVIEW]
-        self.resource.get_all.return_value = ALL_ENCLOSURES
+        self.resource.data = ENCLOSURE_FROM_ONEVIEW
+        self.resource.get_by_hostname.return_value = self.resource
 
         self.mock_ansible_module.params = PARAMS_FOR_PRESENT_NO_HOSTNAME
 
@@ -319,9 +332,12 @@ class TestEnclosureModule(OneViewBaseTest):
         updated_data = ENCLOSURE_FROM_ONEVIEW.copy()
         updated_data['name'] = 'Test-Enclosure-Renamed'
 
-        self.resource.get_by.return_value = [ENCLOSURE_FROM_ONEVIEW]
-        self.resource.get_all.return_value = ALL_ENCLOSURES
-        self.resource.patch.return_value = updated_data
+        self.resource.data = ENCLOSURE_FROM_ONEVIEW
+        self.resource.get_by_hostname.return_value = self.resource
+
+        updated_obj = self.resource.copy()
+        updated_obj.data = updated_data
+        self.resource.patch.return_value = updated_obj
 
         self.mock_ansible_module.params = PARAMS_WITH_NEW_NAME
 
@@ -330,16 +346,18 @@ class TestEnclosureModule(OneViewBaseTest):
         self.mock_ansible_module.exit_json.assert_called_once_with(
             changed=True,
             msg=EnclosureModule.MSG_UPDATED,
-            ansible_facts=dict(enclosure=updated_data)
+            ansible_facts=dict(enclosure=self.resource.data)
         )
 
     def test_update_when_data_has_new_rack_name(self):
         updated_data = ENCLOSURE_FROM_ONEVIEW.copy()
         updated_data['rackName'] = 'Another-Rack-Name'
 
-        self.resource.get_by.return_value = [ENCLOSURE_FROM_ONEVIEW]
-        self.resource.get_all.return_value = ALL_ENCLOSURES
-        self.resource.patch.return_value = updated_data
+        self.resource.data = ENCLOSURE_FROM_ONEVIEW
+
+        updated_obj = self.resource.copy()
+        updated_obj.data = updated_data
+        self.resource.patch.return_value = updated_obj
 
         self.mock_ansible_module.params = PARAMS_WITH_NEW_RACK_NAME
 
@@ -348,13 +366,17 @@ class TestEnclosureModule(OneViewBaseTest):
         self.mock_ansible_module.exit_json.assert_called_once_with(
             changed=True,
             msg=EnclosureModule.MSG_UPDATED,
-            ansible_facts=dict(enclosure=updated_data)
+            ansible_facts=dict(enclosure=self.resource.data)
         )
 
     def test_replace_name_for_new_enclosure(self):
-        self.resource.get_by.return_value = []
+        self.resource.get_by_name.return_value = []
+        self.resource.get_by_hostname.return_value = []
         self.resource.get_all.return_value = []
-        self.resource.add.return_value = ENCLOSURE_FROM_ONEVIEW
+
+        self.resource.data = ENCLOSURE_FROM_ONEVIEW
+        self.resource.add.return_value = self.resource
+
         self.resource.patch.return_value = []
 
         params_ansible = deepcopy(PARAMS_FOR_PRESENT)
@@ -363,28 +385,27 @@ class TestEnclosureModule(OneViewBaseTest):
 
         EnclosureModule().run()
 
-        self.resource.patch.assert_called_once_with(
-            "/a/path", "replace", "/name", "Encl1-Renamed")
+        self.resource.patch.assert_called_once_with("replace", "/name", "Encl1-Renamed")
 
     def test_replace_name_for_existent_enclosure(self):
-        self.resource.get_by.return_value = [ENCLOSURE_FROM_ONEVIEW]
+        self.resource.data = ENCLOSURE_FROM_ONEVIEW
         self.resource.patch.return_value = []
 
         self.mock_ansible_module.params = PARAMS_WITH_NEW_NAME
 
         EnclosureModule().run()
 
-        self.resource.patch.assert_called_once_with(
-            "/a/path", "replace", "/name", "OneView-Enclosure")
+        self.resource.patch.assert_called_once_with("replace", "/name", "OneView-Enclosure")
 
     def test_replace_rack_name_for_new_enclosure(self):
         updated_data = ENCLOSURE_FROM_ONEVIEW.copy()
         updated_data['rackName'] = 'Another-Rack-Name'
 
-        self.resource.get_by.return_value = []
-        self.resource.get_all.return_value = []
-        self.resource.add.return_value = ENCLOSURE_FROM_ONEVIEW
-        self.resource.patch.return_value = []
+        self.resource.get_by_name.return_value = []
+        self.resource.get_by_hostname.return_value = []
+        self.resource.data = ENCLOSURE_FROM_ONEVIEW
+        self.resource.add.return_value = self.resource
+        self.resource.patch.return_value = updated_data
 
         params_ansible = deepcopy(PARAMS_FOR_PRESENT)
         params_ansible['data']['rackName'] = 'Another-Rack-Name'
@@ -393,10 +414,10 @@ class TestEnclosureModule(OneViewBaseTest):
         EnclosureModule().run()
 
         self.resource.patch.assert_called_once_with(
-            "/a/path", "replace", "/rackName", "Another-Rack-Name")
+            "replace", "/rackName", "Another-Rack-Name")
 
     def test_replace_rack_name_for_existent_enclosure(self):
-        self.resource.get_by.return_value = [ENCLOSURE_FROM_ONEVIEW]
+        self.resource.data = ENCLOSURE_FROM_ONEVIEW
         self.resource.patch.return_value = []
 
         self.mock_ansible_module.params = PARAMS_WITH_NEW_RACK_NAME
@@ -404,10 +425,10 @@ class TestEnclosureModule(OneViewBaseTest):
         EnclosureModule().run()
 
         self.resource.patch.assert_called_once_with(
-            "/a/path", "replace", "/rackName", "Another-Rack-Name")
+            "replace", "/rackName", "Another-Rack-Name")
 
     def test_update_calibrated_max_power_for_existent_enclosure(self):
-        self.resource.get_by.return_value = [ENCLOSURE_FROM_ONEVIEW]
+        self.resource.data = ENCLOSURE_FROM_ONEVIEW
         self.resource.patch.return_value = []
 
         self.mock_ansible_module.params = PARAMS_WITH_CALIBRATED_MAX_POWER
@@ -415,10 +436,10 @@ class TestEnclosureModule(OneViewBaseTest):
         EnclosureModule().run()
 
         self.resource.update_environmental_configuration.assert_called_once_with(
-            "/a/path", {"calibratedMaxPower": 1750})
+            {"calibratedMaxPower": 1750})
 
     def test_should_remove_enclosure(self):
-        self.resource.get_by.return_value = [ENCLOSURE_FROM_ONEVIEW]
+        self.resource.data = ENCLOSURE_FROM_ONEVIEW
 
         self.mock_ansible_module.params = PARAMS_FOR_ABSENT
 
@@ -430,7 +451,7 @@ class TestEnclosureModule(OneViewBaseTest):
         )
 
     def test_should_do_nothing_when_enclosure_not_exist(self):
-        self.resource.get_by.return_value = []
+        self.resource.get_by_name.return_value = []
 
         self.mock_ansible_module.params = PARAMS_FOR_ABSENT
 
@@ -442,7 +463,7 @@ class TestEnclosureModule(OneViewBaseTest):
         )
 
     def test_should_reconfigure_enclosure(self):
-        self.resource.get_by.return_value = [ENCLOSURE_FROM_ONEVIEW]
+        self.resource.data = ENCLOSURE_FROM_ONEVIEW
         self.resource.update_configuration.return_value = ENCLOSURE_FROM_ONEVIEW
 
         self.mock_ansible_module.params = PARAMS_FOR_RECONFIGURED
@@ -455,15 +476,16 @@ class TestEnclosureModule(OneViewBaseTest):
         )
 
     def test_should_fail_when_enclosure_not_exist(self):
-        self.resource.get_by.return_value = []
+        self.resource.get_by_name.return_value = []
 
         self.mock_ansible_module.params = PARAMS_FOR_RECONFIGURED
         EnclosureModule().run()
 
-        self.mock_ansible_module.fail_json.assert_called_once_with(exception=mock.ANY, msg=EnclosureModule.MSG_ENCLOSURE_NOT_FOUND)
+        self.mock_ansible_module.fail_json.assert_called_once_with(exception=mock.ANY,
+                                                                   msg=EnclosureModule.MSG_ENCLOSURE_NOT_FOUND)
 
     def test_should_fail_when_name_is_not_in_data(self):
-        self.resource.get_by.return_value = []
+        self.resource.get_by_name.return_value = []
 
         params = deepcopy(PARAMS_FOR_RECONFIGURED)
         del params['data']['name']
@@ -474,7 +496,7 @@ class TestEnclosureModule(OneViewBaseTest):
         self.mock_ansible_module.fail_json.assert_called_once_with(exception=mock.ANY, msg=EnclosureModule.MSG_ENCLOSURE_NOT_FOUND)
 
     def test_should_refresh_enclosure(self):
-        self.resource.get_by.return_value = [ENCLOSURE_FROM_ONEVIEW]
+        self.resource.data = ENCLOSURE_FROM_ONEVIEW
         self.resource.get.return_value = ENCLOSURE_FROM_ONEVIEW
 
         self.mock_ansible_module.params = PARAMS_FOR_REFRESH
@@ -488,15 +510,16 @@ class TestEnclosureModule(OneViewBaseTest):
         )
 
     def test_should_power_on_appliance_bays(self):
-        self.resource.get_by.return_value = [ENCLOSURE_FROM_ONEVIEW]
-        self.resource.patch.return_value = ENCLOSURE_FROM_ONEVIEW
+        self.resource.data = ENCLOSURE_FROM_ONEVIEW
+        self.resource.patch.return_value = self.resource
 
         self.mock_ansible_module.params = PARAMS_FOR_BAY_POWER_ON
 
         EnclosureModule().run()
 
-        self.resource.patch.assert_called_once_with(
-            ENCLOSURE_FROM_ONEVIEW['uri'], operation='replace', path='/applianceBays/2/power', value='On')
+        self.resource.patch.assert_called_once_with(operation='replace',
+                                                    path='/applianceBays/2/power',
+                                                    value='On')
 
         self.mock_ansible_module.exit_json.assert_called_once_with(
             changed=True,
@@ -505,7 +528,7 @@ class TestEnclosureModule(OneViewBaseTest):
         )
 
     def test_should_not_power_on_when_state_is_already_on(self):
-        self.resource.get_by.return_value = [ENCLOSURE_FROM_ONEVIEW]
+        self.resource.data = ENCLOSURE_FROM_ONEVIEW
 
         params_power_on_do_nothing = deepcopy(PARAMS_FOR_BAY_POWER_ON)
         params_power_on_do_nothing['data']['bayNumber'] = 1
@@ -522,7 +545,7 @@ class TestEnclosureModule(OneViewBaseTest):
         )
 
     def test_should_fail_when_appliance_bay_not_found_power_on(self):
-        self.resource.get_by.return_value = [ENCLOSURE_FROM_ONEVIEW]
+        self.resource.data = ENCLOSURE_FROM_ONEVIEW
 
         params_power_on_not_found_bay = deepcopy(PARAMS_FOR_BAY_POWER_ON)
         params_power_on_not_found_bay['data']['bayNumber'] = 3
@@ -532,11 +555,12 @@ class TestEnclosureModule(OneViewBaseTest):
 
         self.resource.patch.not_been_called()
 
-        self.mock_ansible_module.fail_json.assert_called_once_with(exception=mock.ANY, msg=EnclosureModule.MSG_BAY_NOT_FOUND)
+        self.mock_ansible_module.fail_json.assert_called_once_with(exception=mock.ANY,
+                                                                   msg=EnclosureModule.MSG_BAY_NOT_FOUND)
 
     def test_should_fail_when_there_are_not_appliance_bays_power_on(self):
         enclosure_without_appliance_bays = dict(ENCLOSURE_FROM_ONEVIEW, applianceBays=[])
-        self.resource.get_by.return_value = [enclosure_without_appliance_bays]
+        self.resource.data = enclosure_without_appliance_bays
 
         self.mock_ansible_module.params = PARAMS_FOR_BAY_POWER_ON
 
@@ -544,18 +568,20 @@ class TestEnclosureModule(OneViewBaseTest):
 
         self.resource.patch.not_been_called()
 
-        self.mock_ansible_module.fail_json.assert_called_once_with(exception=mock.ANY, msg=EnclosureModule.MSG_BAY_NOT_FOUND)
+        self.mock_ansible_module.fail_json.assert_called_once_with(exception=mock.ANY,
+                                                                   msg=EnclosureModule.MSG_BAY_NOT_FOUND)
 
     def test_should_turn_on_uid(self):
-        self.resource.get_by.return_value = [ENCLOSURE_FROM_ONEVIEW]
-        self.resource.patch.return_value = ENCLOSURE_FROM_ONEVIEW
+        self.resource.data = ENCLOSURE_FROM_ONEVIEW
+        self.resource.patch.return_value = self.resource
 
         self.mock_ansible_module.params = yaml.load(PARAMS_FOR_UID_ON)
 
         EnclosureModule().run()
 
-        self.resource.patch.assert_called_once_with(
-            ENCLOSURE_FROM_ONEVIEW['uri'], operation='replace', path='/uidState', value='On')
+        self.resource.patch.assert_called_once_with(operation='replace',
+                                                    path='/uidState',
+                                                    value='On')
 
         self.mock_ansible_module.exit_json.assert_called_once_with(
             changed=True,
@@ -565,7 +591,7 @@ class TestEnclosureModule(OneViewBaseTest):
 
     def test_should_not_set_to_on_when_it_is_already_on(self):
         enclosure_uid_on = dict(ENCLOSURE_FROM_ONEVIEW, uidState='On')
-        self.resource.get_by.return_value = [enclosure_uid_on]
+        self.resource.data = enclosure_uid_on
 
         self.mock_ansible_module.params = yaml.load(PARAMS_FOR_UID_ON)
 
@@ -581,16 +607,16 @@ class TestEnclosureModule(OneViewBaseTest):
 
     def test_should_turn_off_uid(self):
         enclosure_uid_on = dict(ENCLOSURE_FROM_ONEVIEW, uidState='On')
-
-        self.resource.get_by.return_value = [enclosure_uid_on]
-        self.resource.patch.return_value = enclosure_uid_on
+        self.resource.data = enclosure_uid_on
+        self.resource.patch.return_value = self.resource
 
         self.mock_ansible_module.params = yaml.load(PARAMS_FOR_UID_OFF)
 
         EnclosureModule().run()
 
-        self.resource.patch.assert_called_once_with(
-            ENCLOSURE_FROM_ONEVIEW['uri'], operation='replace', path='/uidState', value='Off')
+        self.resource.patch.assert_called_once_with(operation='replace',
+                                                    path='/uidState',
+                                                    value='Off')
 
         self.mock_ansible_module.exit_json.assert_called_once_with(
             changed=True,
@@ -599,7 +625,7 @@ class TestEnclosureModule(OneViewBaseTest):
         )
 
     def test_should_not_set_to_off_when_it_is_already_off(self):
-        self.resource.get_by.return_value = [ENCLOSURE_FROM_ONEVIEW]
+        self.resource.data = ENCLOSURE_FROM_ONEVIEW
 
         self.mock_ansible_module.params = yaml.load(PARAMS_FOR_UID_OFF)
 
@@ -614,15 +640,16 @@ class TestEnclosureModule(OneViewBaseTest):
         )
 
     def test_should_turn_on_uid_manager_bay(self):
-        self.resource.get_by.return_value = [ENCLOSURE_FROM_ONEVIEW]
-        self.resource.patch.return_value = ENCLOSURE_FROM_ONEVIEW
+        self.resource.data = ENCLOSURE_FROM_ONEVIEW
+        self.resource.patch.return_value = self.resource
 
         self.mock_ansible_module.params = yaml.load(PARAMS_FOR_MANAGER_BAY_UID_ON)
 
         EnclosureModule().run()
 
-        self.resource.patch.assert_called_once_with(
-            ENCLOSURE_FROM_ONEVIEW['uri'], operation='replace', path='/managerBays/2/uidState', value='On')
+        self.resource.patch.assert_called_once_with(operation='replace',
+                                                    path='/managerBays/2/uidState',
+                                                    value='On')
 
         self.mock_ansible_module.exit_json.assert_called_once_with(
             changed=True,
@@ -631,7 +658,7 @@ class TestEnclosureModule(OneViewBaseTest):
         )
 
     def test_should_not_set_to_on_when_state_already_on(self):
-        self.resource.get_by.return_value = [ENCLOSURE_FROM_ONEVIEW]
+        self.resource.data = ENCLOSURE_FROM_ONEVIEW
 
         params_manager_bay_uid = yaml.load(PARAMS_FOR_MANAGER_BAY_UID_ON)
         params_manager_bay_uid['data']['bayNumber'] = '1'
@@ -649,7 +676,7 @@ class TestEnclosureModule(OneViewBaseTest):
         )
 
     def test_should_fail_when_manager_bay_not_found(self):
-        self.resource.get_by.return_value = [ENCLOSURE_FROM_ONEVIEW]
+        self.resource.data = ENCLOSURE_FROM_ONEVIEW
 
         params_power_on_not_found_bay = yaml.load(PARAMS_FOR_MANAGER_BAY_UID_ON)
         params_power_on_not_found_bay['data']['bayNumber'] = 3
@@ -659,11 +686,12 @@ class TestEnclosureModule(OneViewBaseTest):
 
         self.resource.patch.not_been_called()
 
-        self.mock_ansible_module.fail_json.assert_called_once_with(exception=mock.ANY, msg=EnclosureModule.MSG_BAY_NOT_FOUND)
+        self.mock_ansible_module.fail_json.assert_called_once_with(exception=mock.ANY,
+                                                                   msg=EnclosureModule.MSG_BAY_NOT_FOUND)
 
     def test_should_fail_when_there_are_not_manager_bays_uid_on(self):
         enclosure_without_appliance_bays = dict(ENCLOSURE_FROM_ONEVIEW, managerBays=[])
-        self.resource.get_by.return_value = [enclosure_without_appliance_bays]
+        self.resource.data = enclosure_without_appliance_bays
 
         self.mock_ansible_module.params = yaml.load(PARAMS_FOR_MANAGER_BAY_UID_ON)
 
@@ -671,18 +699,20 @@ class TestEnclosureModule(OneViewBaseTest):
 
         self.resource.patch.not_been_called()
 
-        self.mock_ansible_module.fail_json.assert_called_once_with(exception=mock.ANY, msg=EnclosureModule.MSG_BAY_NOT_FOUND)
+        self.mock_ansible_module.fail_json.assert_called_once_with(exception=mock.ANY,
+                                                                   msg=EnclosureModule.MSG_BAY_NOT_FOUND)
 
     def test_should_turn_off_uid_manager_bay(self):
-        self.resource.get_by.return_value = [ENCLOSURE_FROM_ONEVIEW]
-        self.resource.patch.return_value = ENCLOSURE_FROM_ONEVIEW
+        self.resource.data = ENCLOSURE_FROM_ONEVIEW
+        self.resource.patch.return_value = self.resource
 
         self.mock_ansible_module.params = yaml.load(PARAMS_FOR_MANAGER_BAY_UID_OFF)
 
         EnclosureModule().run()
 
-        self.resource.patch.assert_called_once_with(
-            ENCLOSURE_FROM_ONEVIEW['uri'], operation='replace', path='/managerBays/1/uidState', value='Off')
+        self.resource.patch.assert_called_once_with(operation='replace',
+                                                    path='/managerBays/1/uidState',
+                                                    value='Off')
 
         self.mock_ansible_module.exit_json.assert_called_once_with(
             changed=True,
@@ -691,7 +721,7 @@ class TestEnclosureModule(OneViewBaseTest):
         )
 
     def test_should_not_set_to_off_when_state_already_off(self):
-        self.resource.get_by.return_value = [ENCLOSURE_FROM_ONEVIEW]
+        self.resource.data = ENCLOSURE_FROM_ONEVIEW
 
         params_manager_bay_uid = yaml.load(PARAMS_FOR_MANAGER_BAY_UID_OFF)
         params_manager_bay_uid['data']['bayNumber'] = '2'
@@ -709,7 +739,7 @@ class TestEnclosureModule(OneViewBaseTest):
         )
 
     def test_should_fail_when_manager_bay_not_found_uid_off(self):
-        self.resource.get_by.return_value = [ENCLOSURE_FROM_ONEVIEW]
+        self.resource.data = ENCLOSURE_FROM_ONEVIEW
 
         params_power_on_not_found_bay = yaml.load(PARAMS_FOR_MANAGER_BAY_UID_OFF)
         params_power_on_not_found_bay['data']['bayNumber'] = 3
@@ -719,11 +749,12 @@ class TestEnclosureModule(OneViewBaseTest):
 
         self.resource.patch.not_been_called()
 
-        self.mock_ansible_module.fail_json.assert_called_once_with(exception=mock.ANY, msg=EnclosureModule.MSG_BAY_NOT_FOUND)
+        self.mock_ansible_module.fail_json.assert_called_once_with(exception=mock.ANY,
+                                                                   msg=EnclosureModule.MSG_BAY_NOT_FOUND)
 
     def test_should_fail_when_there_are_not_manager_bays_uid_off(self):
         enclosure_without_appliance_bays = dict(ENCLOSURE_FROM_ONEVIEW, managerBays=[])
-        self.resource.get_by.return_value = [enclosure_without_appliance_bays]
+        self.resource.data = enclosure_without_appliance_bays
 
         self.mock_ansible_module.params = yaml.load(PARAMS_FOR_MANAGER_BAY_UID_OFF)
 
@@ -731,18 +762,23 @@ class TestEnclosureModule(OneViewBaseTest):
 
         self.resource.patch.not_been_called()
 
-        self.mock_ansible_module.fail_json.assert_called_once_with(exception=mock.ANY, msg=EnclosureModule.MSG_BAY_NOT_FOUND)
+        self.mock_ansible_module.fail_json.assert_called_once_with(exception=mock.ANY,
+                                                                   msg=EnclosureModule.MSG_BAY_NOT_FOUND)
 
     def test_should_perform_an_e_fuse_manager_bay(self):
-        self.resource.get_by.return_value = [ENCLOSURE_FROM_ONEVIEW]
-        self.resource.patch.return_value = ENCLOSURE_FROM_ONEVIEW
+        self.resource.data = ENCLOSURE_FROM_ONEVIEW
+
+        updated_resource = self.resource.copy()
+        updated_resource.data = ENCLOSURE_FROM_ONEVIEW
+        self.resource.patch.return_value = updated_resource
 
         self.mock_ansible_module.params = yaml.load(PARAMS_FOR_MANAGER_BAY_POWER_STATE_E_FUSE)
 
         EnclosureModule().run()
 
-        self.resource.patch.assert_called_once_with(
-            ENCLOSURE_FROM_ONEVIEW['uri'], operation='replace', path='/managerBays/1/bayPowerState', value='E-Fuse')
+        self.resource.patch.assert_called_once_with(operation='replace',
+                                                    path='/managerBays/1/bayPowerState',
+                                                    value='E-Fuse')
 
         self.mock_ansible_module.exit_json.assert_called_once_with(
             changed=True,
@@ -751,7 +787,7 @@ class TestEnclosureModule(OneViewBaseTest):
         )
 
     def test_should_fail_when_manager_bay_not_found_e_fuse(self):
-        self.resource.get_by.return_value = [ENCLOSURE_FROM_ONEVIEW]
+        self.resource.data = ENCLOSURE_FROM_ONEVIEW
 
         params_power_on_not_found_bay = yaml.load(PARAMS_FOR_MANAGER_BAY_POWER_STATE_E_FUSE)
         params_power_on_not_found_bay['data']['bayNumber'] = 3
@@ -761,11 +797,12 @@ class TestEnclosureModule(OneViewBaseTest):
 
         self.resource.patch.not_been_called()
 
-        self.mock_ansible_module.fail_json.assert_called_once_with(exception=mock.ANY, msg=EnclosureModule.MSG_BAY_NOT_FOUND)
+        self.mock_ansible_module.fail_json.assert_called_once_with(exception=mock.ANY,
+                                                                   msg=EnclosureModule.MSG_BAY_NOT_FOUND)
 
     def test_should_fail_when_there_are_not_manager_bays_e_fuse(self):
         enclosure_without_appliance_bays = dict(ENCLOSURE_FROM_ONEVIEW, managerBays=[])
-        self.resource.get_by.return_value = [enclosure_without_appliance_bays]
+        self.resource.data = enclosure_without_appliance_bays
 
         self.mock_ansible_module.params = yaml.load(PARAMS_FOR_MANAGER_BAY_POWER_STATE_E_FUSE)
 
@@ -773,18 +810,20 @@ class TestEnclosureModule(OneViewBaseTest):
 
         self.resource.patch.not_been_called()
 
-        self.mock_ansible_module.fail_json.assert_called_once_with(exception=mock.ANY, msg=EnclosureModule.MSG_BAY_NOT_FOUND)
+        self.mock_ansible_module.fail_json.assert_called_once_with(exception=mock.ANY,
+                                                                   msg=EnclosureModule.MSG_BAY_NOT_FOUND)
 
     def test_should_reset_manager_bay(self):
-        self.resource.get_by.return_value = [ENCLOSURE_FROM_ONEVIEW]
-        self.resource.patch.return_value = ENCLOSURE_FROM_ONEVIEW
+        self.resource.data = ENCLOSURE_FROM_ONEVIEW
+        self.resource.patch.return_value = self.resource
 
         self.mock_ansible_module.params = yaml.load(PARAMS_FOR_MANAGER_BAY_POWER_STATE_RESET)
 
         EnclosureModule().run()
 
-        self.resource.patch.assert_called_once_with(
-            ENCLOSURE_FROM_ONEVIEW['uri'], operation='replace', path='/managerBays/1/bayPowerState', value='Reset')
+        self.resource.patch.assert_called_once_with(operation='replace',
+                                                    path='/managerBays/1/bayPowerState',
+                                                    value='Reset')
 
         self.mock_ansible_module.exit_json.assert_called_once_with(
             changed=True,
@@ -793,7 +832,7 @@ class TestEnclosureModule(OneViewBaseTest):
         )
 
     def test_should_fail_when_manager_bay_not_found_power_reset(self):
-        self.resource.get_by.return_value = [ENCLOSURE_FROM_ONEVIEW]
+        self.resource.data = ENCLOSURE_FROM_ONEVIEW
 
         params_power_on_not_found_bay = yaml.load(PARAMS_FOR_MANAGER_BAY_POWER_STATE_RESET)
         params_power_on_not_found_bay['data']['bayNumber'] = 3
@@ -803,11 +842,12 @@ class TestEnclosureModule(OneViewBaseTest):
 
         self.resource.patch.not_been_called()
 
-        self.mock_ansible_module.fail_json.assert_called_once_with(exception=mock.ANY, msg=EnclosureModule.MSG_BAY_NOT_FOUND)
+        self.mock_ansible_module.fail_json.assert_called_once_with(exception=mock.ANY,
+                                                                   msg=EnclosureModule.MSG_BAY_NOT_FOUND)
 
     def test_should_fail_when_there_are_not_manager_bays_reset(self):
         enclosure_without_appliance_bays = dict(ENCLOSURE_FROM_ONEVIEW, managerBays=[])
-        self.resource.get_by.return_value = [enclosure_without_appliance_bays]
+        self.resource.data = enclosure_without_appliance_bays
 
         self.mock_ansible_module.params = yaml.load(PARAMS_FOR_MANAGER_BAY_POWER_STATE_RESET)
 
@@ -815,18 +855,20 @@ class TestEnclosureModule(OneViewBaseTest):
 
         self.resource.patch.not_been_called()
 
-        self.mock_ansible_module.fail_json.assert_called_once_with(exception=mock.ANY, msg=EnclosureModule.MSG_BAY_NOT_FOUND)
+        self.mock_ansible_module.fail_json.assert_called_once_with(exception=mock.ANY,
+                                                                   msg=EnclosureModule.MSG_BAY_NOT_FOUND)
 
     def test_should_perform_an_e_fuse_appliance_bay(self):
-        self.resource.get_by.return_value = [ENCLOSURE_FROM_ONEVIEW]
-        self.resource.patch.return_value = ENCLOSURE_FROM_ONEVIEW
+        self.resource.data = ENCLOSURE_FROM_ONEVIEW
+        self.resource.patch.return_value = self.resource
 
         self.mock_ansible_module.params = yaml.load(PARAMS_FOR_APPLIANCE_BAY_POWER_STATE_E_FUSE)
 
         EnclosureModule().run()
 
-        self.resource.patch.assert_called_once_with(
-            ENCLOSURE_FROM_ONEVIEW['uri'], operation='replace', path='/applianceBays/1/bayPowerState', value='E-Fuse')
+        self.resource.patch.assert_called_once_with(operation='replace',
+                                                    path='/applianceBays/1/bayPowerState',
+                                                    value='E-Fuse')
 
         self.mock_ansible_module.exit_json.assert_called_once_with(
             changed=True,
@@ -835,7 +877,7 @@ class TestEnclosureModule(OneViewBaseTest):
         )
 
     def test_should_fail_when_appliance_bay_not_found_appliance_bay(self):
-        self.resource.get_by.return_value = [ENCLOSURE_FROM_ONEVIEW]
+        self.resource.data = ENCLOSURE_FROM_ONEVIEW
 
         params_power_on_not_found_bay = yaml.load(PARAMS_FOR_APPLIANCE_BAY_POWER_STATE_E_FUSE)
         params_power_on_not_found_bay['data']['bayNumber'] = 3
@@ -845,11 +887,12 @@ class TestEnclosureModule(OneViewBaseTest):
 
         self.resource.patch.not_been_called()
 
-        self.mock_ansible_module.fail_json.assert_called_once_with(exception=mock.ANY, msg=EnclosureModule.MSG_BAY_NOT_FOUND)
+        self.mock_ansible_module.fail_json.assert_called_once_with(exception=mock.ANY,
+                                                                   msg=EnclosureModule.MSG_BAY_NOT_FOUND)
 
     def test_should_fail_when_there_are_not_appliance_bays_e_fuse(self):
         enclosure_without_appliance_bays = dict(ENCLOSURE_FROM_ONEVIEW, applianceBays=[])
-        self.resource.get_by.return_value = [enclosure_without_appliance_bays]
+        self.resource.data = enclosure_without_appliance_bays
 
         self.mock_ansible_module.params = yaml.load(PARAMS_FOR_APPLIANCE_BAY_POWER_STATE_E_FUSE)
 
@@ -857,18 +900,20 @@ class TestEnclosureModule(OneViewBaseTest):
 
         self.resource.patch.not_been_called()
 
-        self.mock_ansible_module.fail_json.assert_called_once_with(exception=mock.ANY, msg=EnclosureModule.MSG_BAY_NOT_FOUND)
+        self.mock_ansible_module.fail_json.assert_called_once_with(exception=mock.ANY,
+                                                                   msg=EnclosureModule.MSG_BAY_NOT_FOUND)
 
     def test_should_perform_an_e_fuse_device_bay(self):
-        self.resource.get_by.return_value = [ENCLOSURE_FROM_ONEVIEW]
-        self.resource.patch.return_value = ENCLOSURE_FROM_ONEVIEW
+        self.resource.data = ENCLOSURE_FROM_ONEVIEW
+        self.resource.patch.return_value = self.resource
 
         self.mock_ansible_module.params = yaml.load(PARAMS_FOR_DEVICE_BAY_POWER_STATE_E_FUSE)
 
         EnclosureModule().run()
 
-        self.resource.patch.assert_called_once_with(
-            ENCLOSURE_FROM_ONEVIEW['uri'], operation='replace', path='/deviceBays/1/bayPowerState', value='E-Fuse')
+        self.resource.patch.assert_called_once_with(operation='replace',
+                                                    path='/deviceBays/1/bayPowerState',
+                                                    value='E-Fuse')
 
         self.mock_ansible_module.exit_json.assert_called_once_with(
             changed=True,
@@ -877,7 +922,7 @@ class TestEnclosureModule(OneViewBaseTest):
         )
 
     def test_should_fail_when_device_bay_not_found_e_fuse(self):
-        self.resource.get_by.return_value = [ENCLOSURE_FROM_ONEVIEW]
+        self.resource.data = ENCLOSURE_FROM_ONEVIEW
 
         params_power_on_not_found_bay = yaml.load(PARAMS_FOR_DEVICE_BAY_POWER_STATE_E_FUSE)
         params_power_on_not_found_bay['data']['bayNumber'] = 3
@@ -887,11 +932,12 @@ class TestEnclosureModule(OneViewBaseTest):
 
         self.resource.patch.not_been_called()
 
-        self.mock_ansible_module.fail_json.assert_called_once_with(exception=mock.ANY, msg=EnclosureModule.MSG_BAY_NOT_FOUND)
+        self.mock_ansible_module.fail_json.assert_called_once_with(exception=mock.ANY,
+                                                                   msg=EnclosureModule.MSG_BAY_NOT_FOUND)
 
     def test_should_fail_when_there_are_not_device_bays_e_fuse(self):
         enclosure_without_appliance_bays = dict(ENCLOSURE_FROM_ONEVIEW, deviceBays=[])
-        self.resource.get_by.return_value = [enclosure_without_appliance_bays]
+        self.resource.data = enclosure_without_appliance_bays
 
         self.mock_ansible_module.params = yaml.load(PARAMS_FOR_DEVICE_BAY_POWER_STATE_E_FUSE)
 
@@ -899,18 +945,20 @@ class TestEnclosureModule(OneViewBaseTest):
 
         self.resource.patch.not_been_called()
 
-        self.mock_ansible_module.fail_json.assert_called_once_with(exception=mock.ANY, msg=EnclosureModule.MSG_BAY_NOT_FOUND)
+        self.mock_ansible_module.fail_json.assert_called_once_with(exception=mock.ANY,
+                                                                   msg=EnclosureModule.MSG_BAY_NOT_FOUND)
 
     def test_should_reset_device_bay(self):
-        self.resource.get_by.return_value = [ENCLOSURE_FROM_ONEVIEW]
-        self.resource.patch.return_value = ENCLOSURE_FROM_ONEVIEW
+        self.resource.data = ENCLOSURE_FROM_ONEVIEW
+        self.resource.patch.return_value = self.resource
 
         self.mock_ansible_module.params = yaml.load(PARAMS_FOR_DEVICE_BAY_POWER_STATE_RESET)
 
         EnclosureModule().run()
 
-        self.resource.patch.assert_called_once_with(
-            ENCLOSURE_FROM_ONEVIEW['uri'], operation='replace', path='/deviceBays/1/bayPowerState', value='Reset')
+        self.resource.patch.assert_called_once_with(operation='replace',
+                                                    path='/deviceBays/1/bayPowerState',
+                                                    value='Reset')
 
         self.mock_ansible_module.exit_json.assert_called_once_with(
             changed=True,
@@ -919,7 +967,7 @@ class TestEnclosureModule(OneViewBaseTest):
         )
 
     def test_should_fail_when_device_bay_not_found_reset(self):
-        self.resource.get_by.return_value = [ENCLOSURE_FROM_ONEVIEW]
+        self.resource.data = ENCLOSURE_FROM_ONEVIEW
 
         params_power_on_not_found_bay = yaml.load(PARAMS_FOR_DEVICE_BAY_POWER_STATE_RESET)
         params_power_on_not_found_bay['data']['bayNumber'] = 3
@@ -929,11 +977,12 @@ class TestEnclosureModule(OneViewBaseTest):
 
         self.resource.patch.not_been_called()
 
-        self.mock_ansible_module.fail_json.assert_called_once_with(exception=mock.ANY, msg=EnclosureModule.MSG_BAY_NOT_FOUND)
+        self.mock_ansible_module.fail_json.assert_called_once_with(exception=mock.ANY,
+                                                                   msg=EnclosureModule.MSG_BAY_NOT_FOUND)
 
     def test_should_fail_when_there_are_not_device_bays_reset(self):
         enclosure_without_appliance_bays = dict(ENCLOSURE_FROM_ONEVIEW, deviceBays=[])
-        self.resource.get_by.return_value = [enclosure_without_appliance_bays]
+        self.resource.data = enclosure_without_appliance_bays
 
         self.mock_ansible_module.params = yaml.load(PARAMS_FOR_DEVICE_BAY_POWER_STATE_RESET)
 
@@ -941,19 +990,20 @@ class TestEnclosureModule(OneViewBaseTest):
 
         self.resource.patch.not_been_called()
 
-        self.mock_ansible_module.fail_json.assert_called_once_with(exception=mock.ANY, msg=EnclosureModule.MSG_BAY_NOT_FOUND)
+        self.mock_ansible_module.fail_json.assert_called_once_with(exception=mock.ANY,
+                                                                   msg=EnclosureModule.MSG_BAY_NOT_FOUND)
 
     def test_should_perform_an_e_fuse_interconnect(self):
-        self.resource.get_by.return_value = [ENCLOSURE_FROM_ONEVIEW]
-        self.resource.patch.return_value = ENCLOSURE_FROM_ONEVIEW
+        self.resource.data = ENCLOSURE_FROM_ONEVIEW
+        self.resource.patch.return_value = self.resource
 
         self.mock_ansible_module.params = yaml.load(PARAMS_FOR_INTERCONNECT_BAY_POWER_STATE_E_FUSE)
 
         EnclosureModule().run()
 
-        self.resource.patch.assert_called_once_with(
-            ENCLOSURE_FROM_ONEVIEW['uri'], operation='replace', path='/interconnectBays/2/bayPowerState',
-            value='E-Fuse')
+        self.resource.patch.assert_called_once_with(operation='replace',
+                                                    path='/interconnectBays/2/bayPowerState',
+                                                    value='E-Fuse')
 
         self.mock_ansible_module.exit_json.assert_called_once_with(
             changed=True,
@@ -962,7 +1012,7 @@ class TestEnclosureModule(OneViewBaseTest):
         )
 
     def test_should_fail_when_interconnect_bay_not_found_e_fuse(self):
-        self.resource.get_by.return_value = [ENCLOSURE_FROM_ONEVIEW]
+        self.resource.data = ENCLOSURE_FROM_ONEVIEW
 
         params_power_on_not_found_bay = yaml.load(PARAMS_FOR_INTERCONNECT_BAY_POWER_STATE_E_FUSE)
         params_power_on_not_found_bay['data']['bayNumber'] = 3
@@ -976,7 +1026,7 @@ class TestEnclosureModule(OneViewBaseTest):
 
     def test_should_fail_when_there_are_not_interconnect_bays_e_fuse(self):
         enclosure_without_appliance_bays = dict(ENCLOSURE_FROM_ONEVIEW, interconnectBays=[])
-        self.resource.get_by.return_value = [enclosure_without_appliance_bays]
+        self.resource.data = enclosure_without_appliance_bays
 
         self.mock_ansible_module.params = yaml.load(PARAMS_FOR_INTERCONNECT_BAY_POWER_STATE_E_FUSE)
 
@@ -987,15 +1037,16 @@ class TestEnclosureModule(OneViewBaseTest):
         self.mock_ansible_module.fail_json.assert_called_once_with(exception=mock.ANY, msg=EnclosureModule.MSG_BAY_NOT_FOUND)
 
     def test_should_remove_ipv4_device_bays(self):
-        self.resource.get_by.return_value = [ENCLOSURE_FROM_ONEVIEW]
-        self.resource.patch.return_value = ENCLOSURE_FROM_ONEVIEW
+        self.resource.data = ENCLOSURE_FROM_ONEVIEW
+        self.resource.patch.return_value = self.resource
 
         self.mock_ansible_module.params = yaml.load(PARAMS_FOR_DEVICE_BAY_IPV4_RELEASE)
 
         EnclosureModule().run()
 
-        self.resource.patch.assert_called_once_with(
-            ENCLOSURE_FROM_ONEVIEW['uri'], operation='remove', path='/deviceBays/1/ipv4Setting', value='')
+        self.resource.patch.assert_called_once_with(operation='remove',
+                                                    path='/deviceBays/1/ipv4Setting',
+                                                    value='')
 
         self.mock_ansible_module.exit_json.assert_called_once_with(
             changed=True,
@@ -1004,15 +1055,17 @@ class TestEnclosureModule(OneViewBaseTest):
         )
 
     def test_should_remove_ipv4_interconnect_bays(self):
-        self.resource.get_by.return_value = [ENCLOSURE_FROM_ONEVIEW]
-        self.resource.patch.return_value = ENCLOSURE_FROM_ONEVIEW
+        self.resource.data = ENCLOSURE_FROM_ONEVIEW
+
+        self.resource.patch.return_value = self.resource
 
         self.mock_ansible_module.params = yaml.load(PARAMS_FOR_INTERCONNECT_BAY_IPV4_RELEASE)
 
         EnclosureModule().run()
 
-        self.resource.patch.assert_called_once_with(
-            ENCLOSURE_FROM_ONEVIEW['uri'], operation='remove', path='/interconnectBays/1/ipv4Setting', value='')
+        self.resource.patch.assert_called_once_with(operation='remove',
+                                                    path='/interconnectBays/1/ipv4Setting',
+                                                    value='')
 
         self.mock_ansible_module.exit_json.assert_called_once_with(
             changed=True,
@@ -1021,7 +1074,7 @@ class TestEnclosureModule(OneViewBaseTest):
         )
 
     def test_should_fail_when_device_bay_not_found_ipv4_release(self):
-        self.resource.get_by.return_value = [ENCLOSURE_FROM_ONEVIEW]
+        self.resource.data = ENCLOSURE_FROM_ONEVIEW
 
         params_power_on_not_found_bay = yaml.load(PARAMS_FOR_DEVICE_BAY_IPV4_RELEASE)
         params_power_on_not_found_bay['data']['bayNumber'] = 3
@@ -1035,7 +1088,7 @@ class TestEnclosureModule(OneViewBaseTest):
 
     def test_should_fail_when_there_are_not_device_bays_ipv4_release(self):
         enclosure_without_appliance_bays = dict(ENCLOSURE_FROM_ONEVIEW, deviceBays=[])
-        self.resource.get_by.return_value = [enclosure_without_appliance_bays]
+        self.resource.data = enclosure_without_appliance_bays
 
         self.mock_ansible_module.params = yaml.load(PARAMS_FOR_DEVICE_BAY_IPV4_RELEASE)
 
@@ -1046,7 +1099,7 @@ class TestEnclosureModule(OneViewBaseTest):
         self.mock_ansible_module.fail_json.assert_called_once_with(exception=mock.ANY, msg=EnclosureModule.MSG_BAY_NOT_FOUND)
 
     def test_should_fail_when_interconnect_bay_not_found_ipv4(self):
-        self.resource.get_by.return_value = [ENCLOSURE_FROM_ONEVIEW]
+        self.resource.data = ENCLOSURE_FROM_ONEVIEW
 
         params_power_on_not_found_bay = yaml.load(PARAMS_FOR_DEVICE_BAY_IPV4_RELEASE)
         params_power_on_not_found_bay['data']['bayNumber'] = 3
@@ -1060,7 +1113,7 @@ class TestEnclosureModule(OneViewBaseTest):
 
     def test_should_fail_when_there_are_not_interconnect_bays_ipv4(self):
         enclosure_without_appliance_bays = dict(ENCLOSURE_FROM_ONEVIEW, interconnectBays=[])
-        self.resource.get_by.return_value = [enclosure_without_appliance_bays]
+        self.resource.data = enclosure_without_appliance_bays
 
         self.mock_ansible_module.params = yaml.load(PARAMS_FOR_INTERCONNECT_BAY_IPV4_RELEASE)
 
@@ -1071,15 +1124,17 @@ class TestEnclosureModule(OneViewBaseTest):
         self.mock_ansible_module.fail_json.assert_called_once_with(exception=mock.ANY, msg=EnclosureModule.MSG_BAY_NOT_FOUND)
 
     def test_should_set_state(self):
-        self.resource.get_by.return_value = [ENCLOSURE_FROM_ONEVIEW]
-        self.resource.patch.return_value = ENCLOSURE_FROM_ONEVIEW
+        self.resource.data = ENCLOSURE_FROM_ONEVIEW
+
+        self.resource.patch.return_value = self.resource
 
         self.mock_ansible_module.params = yaml.load(PARAMS_FOR_DATA_COL_SET)
 
         EnclosureModule().run()
 
-        self.resource.patch.assert_called_once_with(ENCLOSURE_FROM_ONEVIEW['uri'], operation='replace',
-                                                    path='/supportDataCollectionState', value='PendingCollection')
+        self.resource.patch.assert_called_once_with(operation='replace',
+                                                    path='/supportDataCollectionState',
+                                                    value='PendingCollection')
 
         self.mock_ansible_module.exit_json.assert_called_once_with(
             changed=True,
@@ -1089,7 +1144,7 @@ class TestEnclosureModule(OneViewBaseTest):
 
     def test_should_not_set_state_when_it_is_already_on_desired_state(self):
         enclosure_uid_on = dict(ENCLOSURE_FROM_ONEVIEW, supportDataCollectionState='PendingCollection')
-        self.resource.get_by.return_value = [enclosure_uid_on]
+        self.resource.data = enclosure_uid_on
 
         self.mock_ansible_module.params = yaml.load(PARAMS_FOR_DATA_COL_SET)
 
@@ -1112,19 +1167,23 @@ class TestEnclosureModule(OneViewBaseTest):
         resource_data['uri'] = 'rest/enclosures/fake'
         resource_data['scopeUris'] = []
 
-        self.resource.get_by.return_value = [resource_data]
-        self.resource.patch.return_value = params_to_scope['data']
+        self.resource.data = resource_data
+
+        patch_return = resource_data.copy()
+        patch_return['scopeUris'] = ['test']
+        patch_return_obj = self.resource.copy()
+        patch_return_obj.data = patch_return
+        self.resource.patch.return_value = patch_return_obj
 
         EnclosureModule().run()
 
-        self.resource.patch.assert_called_once_with('rest/enclosures/fake',
-                                                    operation='replace',
+        self.resource.patch.assert_called_once_with(operation='replace',
                                                     path='/scopeUris',
                                                     value=['test'])
 
         self.mock_ansible_module.exit_json.assert_called_once_with(
             changed=True,
-            ansible_facts=dict(enclosure=params_to_scope['data']),
+            ansible_facts=dict(enclosure=patch_return),
             msg=EnclosureModule.MSG_UPDATED
         )
 
@@ -1133,7 +1192,7 @@ class TestEnclosureModule(OneViewBaseTest):
         params_to_scope['data']['scopeUris'] = ['test']
         self.mock_ansible_module.params = params_to_scope
 
-        self.resource.get_by.return_value = [params_to_scope['data']]
+        self.resource.data = params_to_scope['data']
 
         EnclosureModule().run()